+++ src/inmanta/env.py
@@ -147,6 +147,9 @@ class VirtualEnv(object):
                 sys.path.remove(item)
         sys.path[:0] = new_sys_path
 
+        # Also set the python path environment variable for any subprocess
+        os.environ["PYTHONPATH"] = os.pathsep.join(sys.path)
+
     def _parse_line(self, req_line: str) -> Tuple[Optional[str], str]:
         """
         Parse the requirement line
@@ -164,6 +167,10 @@ class VirtualEnv(object):
         return None, req_line
 
     def _gen_requirements_file(self, requirements_list: List[str]) -> str:
+        """Generate a new requirements file based on the requirements list that was built from all the different modules.
+        :param requirements_list:  A list of requirements from all the requirements files in all modules.
+        :return: A string that can be written to a requirements file that pip understands.
+        """
         modules: Dict[str, Any] = {}
         for req in requirements_list:
             parsed_name, req_spec = self._parse_line(req)
@@ -267,7 +274,7 @@ class VirtualEnv(object):
 
     def _set_current_requirements_hash(self, new_hash):
         """
-        Set the current requirements hahs
+        Set the current requirements hash
         """
         path = os.path.join(self.env_path, "requirements.sha1sum")
         with open(path, "w+", encoding="utf-8") as fd:
@@ -282,7 +289,7 @@ class VirtualEnv(object):
             if len(requirements_list) == 0:
                 return
 
-        requirements_list = sorted(self._remove_requirements_present_in_parent_env(requirements_list))
+        requirements_list = sorted(requirements_list)
 
         # hash it
         sha1sum = hashlib.sha1()
@@ -299,27 +306,20 @@ class VirtualEnv(object):
         for x in requirements_list:
             self.__cache_done.add(x)
 
-    def _remove_requirements_present_in_parent_env(self, requirements_list: List[str]) -> List[str]:
-        reqs_to_remove = []
-        packages_installed_in_parent = self.get_package_installed_in_parent_env()
-        for r in requirements_list:
-            # Always install url-based requirements
-            if "://" in r:
-                continue
-            parsed_req = list(pkg_resources.parse_requirements(r))[0]
-            # Package is installed and its version fits the constraint
-            if (
-                parsed_req.project_name in packages_installed_in_parent
-                and packages_installed_in_parent[parsed_req.project_name] in parsed_req
-            ):
-                reqs_to_remove.append(r)
-        return [r for r in requirements_list if r not in reqs_to_remove]
-
     @classmethod
-    def _get_installed_packages(cls, python_interpreter: str) -> Dict[str, str]:
+    def _get_installed_packages(cls, python_interpreter: str, inherit_python_path: bool = True) -> Dict[str, str]:
+        """Return a list of all installed packages in the site-packages of a python interpreter.
+        :param python_interpreter: The python interpreter to get the packages for
+        :param inherit_python_path: Should it inherit the python path from this process or not?
+        :return: A dict with package names as keys and versions as values
+        """
         cmd = [python_interpreter, "-m", "pip", "list", "--format", "json"]
         try:
-            output = subprocess.check_output(cmd, stderr=subprocess.DEVNULL)
+            environment = os.environ.copy()
+            if not inherit_python_path and "PYTHONPATH" in environment:
+                del environment["PYTHONPATH"]
+
+            output = subprocess.check_output(cmd, stderr=subprocess.DEVNULL, env=environment)
         except CalledProcessError as e:
             LOGGER.error("%s: %s", cmd, e.output.decode())
             raise